apollo_infra: move test utils to relevant module
